requirements.txt
scrapy>=1.3.2
scrapy-redis>=0.6.8
pymongo>=3.4.0
selenium>=3.0.2
python-dateutil>=2.6.0
